requirements.txt
# Core dependencies (none required for basic functionality)

# Optional dependencies
pyyaml>=6.0  # For YAML config support

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
pytest-mock>=3.0
black>=23.0
isort>=5.0
mypy>=1.0
flake8>=6.0
